// types.ts
export interface Product {
  category: string;
  price: string;
  stocked: boolean;
  name: string;
}

export interface FilterableProductTableProps {
  products: Product[];
}
export interface FilterableProductTableState {
  filterText: string;
  inStockOnly: boolean;
}

export type ProductTableProps = FilterableProductTableProps & FilterableProductTableState;

export interface SearchBarProps extends FilterableProductTableState {
  onFilterTextChange: (e: any) => void;
  onInStockChange: (e: any) => void;
}

export interface ProductRowProps {
  product: Product;
}

export type ProductCategoryRowProps =  Pick<Product, "category">;